Adding a trained neural network and a framework to run it inside of larlight.
The package FANN contains the neural net, and FANNModule is the class you want to actually run a neural network.  Right now there is one trained network available, cascade_net.net. It is meant to take the output of CPAN's getFANNVector function and return track or shower.

For questions about FANN, feel free to ask me.

-Corey