CMAlgo Updates:
- CMalgoAngleCompat and CMAlgoShortestDist: added option to set minimum number of hits for cluster to be taken in consideration by algorithm.
Automatically set to 0 and if algorithm is run as-is it will work as before.
- CMalgoStartInCone: now it works as intended! Also added an option to extend a cluster's "cone" outward beyond its length. This parameter is
set as a fraction of the cone's length. Algorithm will merge two clusters if the start point of one is in the cone of the other. The cone starts
at the start point and extends towards (and possibly beyond) the end point with an angle equal to the cluster's opening angle.