Updates to nue appearance code.
This update significantly changes the structure of the nue appearance code to allow the user to toggle on and off the inclusion of numus and signal.  So, now a user can build a covariance matrix or event ratio (or event rate) plot that only has nues.

Updates to sensitivity calculations are coming, which should allow the same functionality as well as a big increase in speed when using a covariance matrix.