Changes to Topology Calc, Reprocessing
Changes to the topology calc make it faster and easier to read.  It also does a better job calculating all of the entries for a table of event rates w/ systematics.

Changes to reprocessing were to make the interface standardized no matter what the input file change might be.  I will try to maintain it (and not change the input that often)